// src/api/model/policy.ts
/**
 * Policy Tool Service
 * Policy Tool Service Swagger Specification.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: soultos@protonmail.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AdviceExpressions } from './adviceExpressions';
import { DefaultsType } from './defaultsType';
import { ObligationExpressions } from './obligationExpressions';
import { PolicyIssuer } from './policyIssuer';
import { Rule } from './rule';
import { Target } from './target';

export class Policy {
    /**
    * Description
    */
    'description': string;
    'policyIssuer': PolicyIssuer;
    'policyDefaults': DefaultsType;
    'target': Target;
    /**
    * Combiner parameters and rule combiner parameters and variable definitions.
    */
    'combinerParametersAndRuleCombinerParametersAndVariableDefinitions': Array<Rule>;
    'obligationExpressions': ObligationExpressions;
    'adviceExpressions': AdviceExpressions;
    /**
    * Policy ID
    */
    'policyId': string;
    /**
    * Version
    */
    'version': string;
    /**
    * Rule combining algorithm ID
    */
    'ruleCombiningAlgId': string;
    /**
    * Max delegation depth
    */
    'maxDelegationDepth': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "policyIssuer",
            "baseName": "policyIssuer",
            "type": "PolicyIssuer"
        },
        {
            "name": "policyDefaults",
            "baseName": "policyDefaults",
            "type": "DefaultsType"
        },
        {
            "name": "target",
            "baseName": "target",
            "type": "Target"
        },
        {
            "name": "combinerParametersAndRuleCombinerParametersAndVariableDefinitions",
            "baseName": "combinerParametersAndRuleCombinerParametersAndVariableDefinitions",
            "type": "Array<Rule>"
        },
        {
            "name": "obligationExpressions",
            "baseName": "obligationExpressions",
            "type": "ObligationExpressions"
        },
        {
            "name": "adviceExpressions",
            "baseName": "adviceExpressions",
            "type": "AdviceExpressions"
        },
        {
            "name": "policyId",
            "baseName": "policyId",
            "type": "string"
        },
        {
            "name": "version",
            "baseName": "version",
            "type": "string"
        },
        {
            "name": "ruleCombiningAlgId",
            "baseName": "ruleCombiningAlgId",
            "type": "string"
        },
        {
            "name": "maxDelegationDepth",
            "baseName": "maxDelegationDepth",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return Policy.attributeTypeMap;
    }
}