// src/api/model/rule.ts
/**
 * Policy Tool Service
 * Policy Tool Service Swagger Specification.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: soultos@protonmail.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Target } from './target';

export class Rule {
    /**
    * Description
    */
    'description': string;
    'target': Target;
    /**
    * Condition
    */
    'condition': string;
    /**
    * Obligation expressions
    */
    'obligationExpressions': Array<string>;
    /**
    * Advice expressions
    */
    'adviceExpressions': Array<string>;
    /**
    * Rule ID
    */
    'ruleId': string;
    /**
    * Effect
    */
    'effect': Rule.EffectEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "target",
            "baseName": "target",
            "type": "Target"
        },
        {
            "name": "condition",
            "baseName": "condition",
            "type": "string"
        },
        {
            "name": "obligationExpressions",
            "baseName": "obligationExpressions",
            "type": "Array<string>"
        },
        {
            "name": "adviceExpressions",
            "baseName": "adviceExpressions",
            "type": "Array<string>"
        },
        {
            "name": "ruleId",
            "baseName": "ruleId",
            "type": "string"
        },
        {
            "name": "effect",
            "baseName": "effect",
            "type": "Rule.EffectEnum"
        }    ];

    static getAttributeTypeMap() {
        return Rule.attributeTypeMap;
    }
}

export namespace Rule {
    export enum EffectEnum {
        Permit = <any> 'PERMIT',
        Deny = <any> 'DENY'
    }
}