// src/api/model/match.ts
/**
 * Policy Tool Service
 * Policy Tool Service Swagger Specification.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: soultos@protonmail.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AttributeDesignatorType } from './attributeDesignatorType';
import { AttributeSelectorType } from './attributeSelectorType';
import { AttributeValueType } from './attributeValueType';

export class Match {
    'attributeValue': AttributeValueType;
    'attributeSelector': AttributeSelectorType;
    'attributeDesignator': AttributeDesignatorType;
    /**
    * Match ID
    */
    'matchId': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "attributeValue",
            "baseName": "attributeValue",
            "type": "AttributeValueType"
        },
        {
            "name": "attributeSelector",
            "baseName": "attributeSelector",
            "type": "AttributeSelectorType"
        },
        {
            "name": "attributeDesignator",
            "baseName": "attributeDesignator",
            "type": "AttributeDesignatorType"
        },
        {
            "name": "matchId",
            "baseName": "matchId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Match.attributeTypeMap;
    }
}